feat: ZK CIVC  (#16687)
At last! Add randomness to the UltraOp version of the op queue and
ensure proper handling in Translator. This makes the Translator and
Merge proof actually zero knowledge.

The caveat is, because Translator is zero-knowledge by default, we have
to ensure these changes don't affect the AVM recursive verifier. There,
we only ever accumulate one circuit so we don't want to do an artificial
append merge with offset. As such, I opted for translator not expecting
to process random ops at the end of the op queue when the op queue is
derived from the goblin recursive verifier circuit. We however add the
three random ops at the beginning of the op queue by default as
otherwise we'd have to alter Translator relations based on whether
Goblin is used in AVM or CIVC.

Closes https://github.com/AztecProtocol/barretenberg/issues/1360