refactor: moving `emit_note` func, nuking redundant check (#16889)
As I mentioned
[here](https://github.com/AztecProtocol/aztec-packages/pull/16091#discussion_r2324441829)
makes sense to move the `emit_note` function directly to the
`NoteEmission` struct as that was the only callsite of the function.
This allowed me to nuke the old ugly `assert_note_exists` check in [this
commit
](https://github.com/AztecProtocol/aztec-packages/pull/16889/commits/8570bd3e0d4d1fa04fad2b0e0437b2d84809ce12)as
now the API guarantees that we are emitting a real note. This makes this
change a huge win in my opinion.

Closes https://github.com/AztecProtocol/aztec-packages/issues/8589.