fix: oracle interface change check (#16983)
When I modified oracles in [my other
PR](https://github.com/AztecProtocol/aztec-packages/pull/16435) I
noticed that the oracle change check I implemented a while ago in [this
PR](https://github.com/AztecProtocol/aztec-packages/pull/16435) was not
triggered. It turns out that there was a bug and stringifying the object
resulted in "{}" string because the serializer didn't manage to handle
the value types.

Addressed this by just hashing the oracle method names. I think this is
fine because the oracle return values are generally just an array of
fields and hence we are anyway not really able to track there changes in
types. An advantage also is that we can now drop the
"json-stringify-deterministic" dependency from `pxe` package.