chore: Add e2e test for inactivity slash with consecutive epochs (#16932)
Equivalent to existing inactivity slash test, but requires N consecutive
epochs of inactivity to slash.

This required adding a new config entry `proverNodeEpochProvingDelayMs`
which delays the start of a proving job. Issue was that the sentinel
processes missed attestations with a 2-slot delay (to allow time for all
attestations to have been propagated), but the proof for a given epoch
was landing immediately after it ended, so during
`Sentinel.computeProvenPerformance` not all data was available.