chore(bb): lazy logging (#17164)
While working at C++ simulation optimization (ms granularity) I realized
that logging started to take a good chunk of time. This is especially
true now that builds use asserts (not `NDEBUG`) by default and there is
no "release" mode (and nobody will be using it).

The change in this PR makes the formatting of a log lazy for `debug` and
`vinfo`.