feat: Flag to always reexecute block proposals (#17273)
Adds a flag to always reexecute block proposals. If set, a validator
node will always reexecute, even if not part of the committee, though
they will not attest. If the node is not a validator, they will just log
the result fo the execution.

Note that this does NOT affect p2p propagation, since the reexecution is
done after the attestation is propagated, as it happens on a separate
handler and not in a p2p-registered validator.

To handle reexecutions in a non-validator node, reexecution was moved to
a block-proposal-handler class, which is instantiated instead of a
validator client in non-validators.

This PR also causes validators to reexecute a proposal if they are not
in the committee if there is a slash penalty defined for broadcasting
invalid block proposals. Since this feature is not yet properly tested,
I've disabled the default slash for these offenses for the time being
(they were not working at the moment). See A-57 for more info.

Fixes A-54