refactor!: introducing `ContractSelf` (#17884)
Closes
https://linear.app/aztec-labs/issue/F-21/implement-the-self-pattern

Implementing `ContractSelf` that is injected to all the external
functions as `self`. In this PR I do minimal initial implementation and
for this reason the implementation of `ContractSelf` doesn't handle
much.

I wrote quite detailed migration notes that should make it clear what
this PR achieves if it's not already clear from the issue and this
description.

The juicy bits are in `aztec-nr` dir and the changes to contracts are
just repetitive callsite updates.