fix: stricter checks at config update (#18495)
This pr addresses an issue where it was possible to configure the mana
targets such that it would revert if the full limit was actually used by
a block.

In practice the block builder could just build smaller blocks and avoid
it. But it should not be possible to configure something that makes
little sense, so here we are adding checks at the time of update to
handle it.

Also deletes an unused constant.