fix(avm): functionSelector with empty calladata (#18733)
Fixes a crash where the TS simulator reverts AND the calldata is empty.
It tries to extract `calldata[0]` when calling
`callRequest.functionSelector`.

This just replaces `callRequest.functionSelector` with `fnName`. In the
rare case where we revert with empty calldata the log is a bit redundant
(since contract address is printed twice) but that's fine for now.