feat: add noir tests for events (#19128)
This started out as me wanting to remove the commitment from an event
message, as it is unused, and then realizing that testing this was quite
sad because we had no infra to do it.

We now do.

I did not publicly expose it yet as there are some open questions (what
to do if we get too many events? how do we handle events emitted by
contracts?), but those are more API questions and not related to the
internals. With this we can at least use the capabilities ourselves for
internal testing.