chore: Another deflake attempt at multiple validator sentinel e2e (#19237)
In this test, we were waiting for `SLOT_COUNT` to pass and for the first
node to be elected as proposer at least once. But then, when we
inspected the sentinel history, we filtered only by the first
`SLOT_COUNT` slots. So if the first node was elected as proposer after
`SLOT_COUNT`, its history would be empty, which is the error we were
seeing in failed runs like [this
one](http://ci.aztec-labs.com/a88649af146bf02e).